OLMIS-766: Web service specification/documentation
Add proof-of-concepts demonstrating how IDL specs may be automatically tested and documented. Several approaches are shown, with the intent of later removing redundant ones. The code, as is, isn’t ready for production for two reasons. The first is that there’s a mismatch between our current IDL and endpoints. The former specifies simple response bodies, whereas the later uses HATEOUS. Test code that verifies that the IDL and endpoints are in sync therefore uses non-SDR endpoints added for the sake of the demonstration. The second area in which refinement is necessary relates to the need for IP address configuration. Currently, 192.168.99.100 is hardcoded in several places.

The “test” gradle task will now launch a server and run the tests in UserIntegrationTest against its API. While doing so, it ensures that the requests and responses match those defined within the /test/resources/requisition-service.yaml RAML file. This is the approach I suggest we adopt for testing.

The version of UserIntegrationTest in the org.openlmis.referencedata.resteasy package is configured to use the RestEasy rather than RestAssured library. I think the later is much better suited to our needs, and thus plan to remove the RestEasy-based class.

The new “ramlToSwagger” gradle task is intended to be run as part of a build. It converts the aforementioned requisition-service.yaml RAML file into a Swagger equivalent. The Swagger file may then be accessed via an intuitive and interactive UI served at index.html. Although there are RAML-based equivalents to this Swagger UI, I found them less appealing.

The “mochaRun” gradle task introduces a JavaScript-based approach to API and IDL testing. Specifically, it runs /src/integration-test/javascript/test.js. This works well but relies on a running server. An appeal of moving toward one of the previously described Java-based approaches is that they more naturally fit SpringBoot’s build and test lifecycles. With a Java-based approach, there’s no special work necessary to setup a running server.

Finally, this check-in includes a sample SDR projection (UserWithHomeFacilityProjection). OLMIS-671 further describes the thoughts and findings behind this check-in.