OLMIS-750 Change default schema
Change default schema from public to service name. Hibernate does not automatically generate it, so a SQL file is added to create it (and one for reference data), if it does not already exist. Modify reference data entities to specify reference data schema.

Also add pretty print to JSON output.