OLMIS-933 (and OLMIS-1100 a bit) refactored beginning balance calculation
I've done this small refactor while working on OLMOS-933. Currently
we have quite a few places where calculations are done, and it would
be best to start consolidating them in the requisition domain class.

I've moved the beginning balance one, there are few others which can
be moved as part of OLMIS-1100 - requisition calculation service should
most likely cease to exists, so that I don't have to call any calculations
on it additionaly to calling requisition.submit() for example.