OLMIS-1379 Enforce rights and skip line items on skip period
This commit adds right enforcement to the existing skip requisition
period endpoint. Users must have the REQUISTION_CREATE right in order
to be able to skip a period. The program must also allow skipping the
requisition period. Finally, when skipping a requisition period, all
line items are marked as skipped.