OLMIS-1767: Update IntegrationTest
As eluded to in the last commit’s comments, there's a bug wherein JaVers throws an ENTITY_INSTANCE_WITH_NULL_ID error from within the setUp() method of RequisitionControllerIntegrationTest. As a team, we decided that manually setting this ID is how we'd like to sidestep the problem for now. This commit introduces the change.

It also removes requisition.setId() from the generateRequisition method. Setting the requisition's ID there was harmless but unnecessary.