OLMIS-1762: Service starts using flyway
the service is currently using Hibernate to auto-generate the DDL (schema)
based on Hibernate annotations in the Java code and Spring Boot
configuration properties (application.properties). This is convenient for
developers when coding, since the schema is likely to be changed. However,
this does not make sense in a production environment.