OLMIS-1708: Moved BaseReferenceDataService.findXXX methods
The fulfillment service classes will need methods that are very similar
to methods that are in the BaseReferenceDataService class. Because of
that methods have been moved to BaseCommunicationService class. Also the
AuthService, RequestHelper and RequestParameters classes will be used.

Also because ReferenceDataRetrievalException is more generical it had to
be renamed to have more generic name: DataRetrievalException