OLMIS-1869: Fix Attempted Requisition Status Change Silently Ignored
Invariant requisitionDto properties for update endpoints are removed from RAML. That means even if they are in request they are ignored. That entails client of api does not have to put invariant properties in request so they can be null so I had to remove validation that checks invariants didn't changed.