Improve checking last requisition status query
The application logic that was supposed to check the status
of a last requisition was retrieving a whole requisition object from
the database. This meant that the query needed to include all the line
items for requisitions, status changes and other things that weren't needed.
I've adjusted the query to just retrieve the status what should make the
execution time faster.