OLMIS-2681 Change refresh of requisition permission strings to Flyway migration
Since it can be done in a Flyway migration, we should do that. Remove the initializing code, so that it doesn’t run when refresh-db is set. Add logic to demo data and performance data loading mechanisms to use the migration to refresh the table. Since we need to generate a random UUID in SQL, we load the uuid-ossp extension. We separate the creating of extension into a first migration, so that the second migration can be reused in demo data and performance data load.