OLMIS-3332 Add schedule to do clustering on requisition line items
This will ensure the requisition line items table is clustered, so that when line items are retrieved with a requisition, the database should only need to page once.

Two settings are created, one to enable the scheduled task to perform the clustering, the other to specify a cron expression of when the scheduled task should run. These values are set in Spring’s application properties with defaults (not enabled and run every day at midnight) but these can be overridden.