OLMIS-4113: Populate all available products on initiate
Per new emergency requisition workflow, we shouldn't predefine a list of
products for emergency requisitions, rather we should allow users to
select which ones they want to add. Due to this, we should modify how
initiating emergency requisitions work. Instead of populating the
requisition line items with products from the approved list, we should
add them to the list of available products for the requisition - similar
how it's done for non-full supply.