OLMIS-4861 - Add source and additional options for 'skipped' column
Add source options for skipped column of requisition templates. This is for
an administrator to choose if a requisition template should remember 'skipped'
column values from previous period's regular requisition or not.

An additional option is also introduced in this commit to configure
whether the R&R page should disable skipped line items or hide them.

The demo data has been updated to include the default value for the additional option.
The migration script also includes script that migrates existing implementations.

Requisition page changes related to this configuration are not yet implemented.