OLMIS-4714 Initiate SBR with line items for existing stock cards only
This modifies the current behaviour of initiating stock based requisitions
so it does no longer require stock cards for all facility type approved
products. Instead, the requisition will contain a subset of line items
for those stock cards that exist. This may even result in an empty
requsition (that can be deleted or skipped).