OLMIS-4728 Introduce optimistic locking for requisitions
Added a new version field to the requisition domain that is managed
by Hibernate. Each update the version field is incremented in the database.
In case of a mismatch between the version passed to save and the
version in database, Hibernate throws an excepion indicating
optimistic locking violation. Some endpoints now also set
the eTag header in the response with the version number of the requisition.