OLMIS-6784 Fix Sonar and code coverage issue
While trying to improve test coverage, it was discovered that some classes that use PowerMock would be excluded from the Jacoco test coverage report. This is because PowerMock and Jacoco do not work well together, as PowerMock modifies classes during testing and Jacoco requires classes to be the same when creating reports. This can be fixed by configuring Jacoco to use offline instrumentation, but the level of effort is too much, so the alternative to removing the requirement of PowerMock in JasperReportsViewServiceTest. This is also the case in JasperTemplateServiceTest.